sonar lint bug detection pending

+++ journalApp/src/main/java/com/journalingapp/soham/journalApp/service/JournalEntryService.java
@@ -1,11 +1,16 @@
 package com.journalingapp.soham.journalApp.service;
 
 import com.journalingapp.soham.journalApp.entity.JournalEntry;
+import com.journalingapp.soham.journalApp.entity.User;
 import com.journalingapp.soham.journalApp.repository.JournalEntryRepository;
 import org.bson.types.ObjectId;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Transactional;
 
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 
@@ -15,8 +20,30 @@ public class JournalEntryService {
     @Autowired
     private JournalEntryRepository journalEntryRepository;
 
+    @Autowired
+    private UserService userService ;
+
+
+
+
+
+    @Transactional
+    public void saveEntry(JournalEntry  journalEntry, String userName){
+        try {
+            User user = userService.findByUserName(userName);
+            journalEntry.setDate(LocalDateTime.now()) ;
+            JournalEntry saved = journalEntryRepository.save(journalEntry) ;
+            user.getJournalEntries().add(saved);
+//            user.setUserName(null);
+            userService.saveUser(user);
+        }catch (Exception e){
+            System.out.println(e);
+            throw new RuntimeException("An error occurred while saving the entry",e);
+        }
+    }
+
     public void saveEntry(JournalEntry  journalEntry){
-        journalEntryRepository.save(journalEntry) ;
+            journalEntryRepository.save(journalEntry) ;
     }
 
     public List<JournalEntry> getAll(){
@@ -28,10 +55,27 @@ public class JournalEntryService {
 
     }
 
+    @Transactional
+    public boolean deleteById(ObjectId id, String userName){
+        boolean removed = false;
+        try{
+            User user = userService.findByUserName(userName);
+            removed = user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+            if(removed){
+                userService.saveUser(user);
+                journalEntryRepository.deleteById(id);
+            }
 
-    public void deleteById(ObjectId id ){
-        journalEntryRepository.deleteById(id);
+        }catch(Exception e){
+            System.out.println(e);
+            throw new RuntimeException("An error occurred while deleting the entry",e);
+        }
+        return removed ;
     }
+
+//    public List<JournalEntry> findUserName(String userName){
+//        return new List<JournalEntry>;
+//    }
 
 }
